Adds ACK relays and factors relay method into a shared routine.
