+++ serf/event.go
@@ -2,7 +2,6 @@ package serf
 
 import (
 	"fmt"
-	"math/rand"
 	"net"
 	"sync"
 	"time"
@@ -131,12 +130,12 @@ func (q *Query) Respond(buf []byte) error {
 
 	// Check if we've already responded
 	if q.deadline.IsZero() {
-		return fmt.Errorf("Response already sent")
+		return fmt.Errorf("response already sent")
 	}
 
 	// Ensure we aren't past our response deadline
 	if time.Now().After(q.deadline) {
-		return fmt.Errorf("Response is past the deadline")
+		return fmt.Errorf("response is past the deadline")
 	}
 
 	// Create response
@@ -148,83 +147,28 @@ func (q *Query) Respond(buf []byte) error {
 	}
 
 	// Send a direct response
-	{
-		raw, err := encodeMessage(messageQueryResponseType, &resp)
-		if err != nil {
-			return fmt.Errorf("Failed to format response: %v", err)
-		}
+	raw, err := encodeMessage(messageQueryResponseType, &resp)
+	if err != nil {
+		return fmt.Errorf("failed to format response: %v", err)
+	}
 
-		// Check the size limit
-		if len(raw) > q.serf.config.QueryResponseSizeLimit {
-			return fmt.Errorf("response exceeds limit of %d bytes", q.serf.config.QueryResponseSizeLimit)
-		}
+	// Check the size limit
+	if len(raw) > q.serf.config.QueryResponseSizeLimit {
+		return fmt.Errorf("response exceeds limit of %d bytes", q.serf.config.QueryResponseSizeLimit)
+	}
 
-		addr := net.UDPAddr{IP: q.addr, Port: int(q.port)}
-		if err := q.serf.memberlist.SendTo(&addr, raw); err != nil {
-			return err
-		}
+	// Send the response directly to the originator
+	addr := net.UDPAddr{IP: q.addr, Port: int(q.port)}
+	if err := q.serf.memberlist.SendTo(&addr, raw); err != nil {
+		return err
 	}
 
 	// Relay the response through up to relayFactor other nodes
-	members := q.serf.Members()
-	if len(members) > 2 {
-		addr := net.UDPAddr{IP: q.addr, Port: int(q.port)}
-		raw, err := encodeRelayMessage(messageQueryResponseType, addr, &resp)
-		if err != nil {
-			return fmt.Errorf("Failed to format relayed response: %v", err)
-		}
-
-		// Check the size limit
-		if len(raw) > q.serf.config.QueryResponseSizeLimit {
-			return fmt.Errorf("relayed response exceeds limit of %d bytes", q.serf.config.QueryResponseSizeLimit)
-		}
-
-		relayMembers := kRandomMembers(int(q.relayFactor), members, func(m Member) bool {
-			return m.Status != StatusAlive || m.ProtocolMax < 5 || m.Name == q.serf.LocalMember().Name
-		})
-
-		for _, m := range relayMembers {
-			relayAddr := net.UDPAddr{IP: m.Addr, Port: int(m.Port)}
-			if err := q.serf.memberlist.SendTo(&relayAddr, raw); err != nil {
-				return err
-			}
-		}
+	if err := q.serf.relayResponse(q.relayFactor, addr, &resp); err != nil {
+		return err
 	}
 
 	// Clear the deadline, responses sent
 	q.deadline = time.Time{}
 	return nil
-}
-
-// kRandomMembers selects up to k members from a given list, optionally
-// filtering by the given filterFunc
-func kRandomMembers(k int, members []Member, filterFunc func(Member) bool) []Member {
-	n := len(members)
-	kMembers := make([]Member, 0, k)
-OUTER:
-	// Probe up to 3*n times, with large n this is not necessary
-	// since k << n, but with small n we want search to be
-	// exhaustive
-	for i := 0; i < 3*n && len(kMembers) < k; i++ {
-		// Get random member
-		idx := rand.Intn(n)
-		member := members[idx]
-
-		// Give the filter a shot at it.
-		if filterFunc != nil && filterFunc(member) {
-			continue OUTER
-		}
-
-		// Check if we have this member already
-		for j := 0; j < len(kMembers); j++ {
-			if member.Name == kMembers[j].Name {
-				continue OUTER
-			}
-		}
-
-		// Append the member
-		kMembers = append(kMembers, member)
-	}
-
-	return kMembers
 }
