+++ serf/event_test.go
@@ -1,7 +1,6 @@
 package serf
 
 import (
-	"fmt"
 	"reflect"
 	"testing"
 	"time"
@@ -209,59 +208,4 @@ func TestEventType_String(t *testing.T) {
 		}
 	}()
 	other.String()
-}
-
-func TestKRandomNodes(t *testing.T) {
-	nodes := []Member{}
-	for i := 0; i < 90; i++ {
-		// Half the nodes are in a bad state
-		state := StatusAlive
-		switch i % 3 {
-		case 0:
-			state = StatusAlive
-		case 1:
-			state = StatusFailed
-		case 2:
-			state = StatusLeft
-		}
-		nodes = append(nodes, Member{
-			Name:   fmt.Sprintf("test%d", i),
-			Status: state,
-		})
-	}
-
-	filterFunc := func(m Member) bool {
-		if m.Name == "test0" || m.Status != StatusAlive {
-			return true
-		}
-		return false
-	}
-
-	s1 := kRandomMembers(3, nodes, filterFunc)
-	s2 := kRandomMembers(3, nodes, filterFunc)
-	s3 := kRandomMembers(3, nodes, filterFunc)
-
-	if reflect.DeepEqual(s1, s2) {
-		t.Fatalf("unexpected equal")
-	}
-	if reflect.DeepEqual(s1, s3) {
-		t.Fatalf("unexpected equal")
-	}
-	if reflect.DeepEqual(s2, s3) {
-		t.Fatalf("unexpected equal")
-	}
-
-	for _, s := range [][]Member{s1, s2, s3} {
-		if len(s) != 3 {
-			t.Fatalf("bad len")
-		}
-		for _, m := range s {
-			if m.Name == "test0" {
-				t.Fatalf("Bad name")
-			}
-			if m.Status != StatusAlive {
-				t.Fatalf("Bad state")
-			}
-		}
-	}
 }
