+++ serf/query_test.go
@@ -1,6 +1,8 @@
 package serf
 
 import (
+	"fmt"
+	"reflect"
 	"testing"
 	"time"
 )
@@ -134,5 +136,60 @@ func TestSerf_ShouldProcess(t *testing.T) {
 	}
 	if s1.shouldProcessQuery(filters) {
 		t.Fatalf("expected false")
+	}
+}
+
+func Test_kRandomMembers(t *testing.T) {
+	nodes := []Member{}
+	for i := 0; i < 90; i++ {
+		// Half the nodes are in a bad state
+		state := StatusAlive
+		switch i % 3 {
+		case 0:
+			state = StatusAlive
+		case 1:
+			state = StatusFailed
+		case 2:
+			state = StatusLeft
+		}
+		nodes = append(nodes, Member{
+			Name:   fmt.Sprintf("test%d", i),
+			Status: state,
+		})
+	}
+
+	filterFunc := func(m Member) bool {
+		if m.Name == "test0" || m.Status != StatusAlive {
+			return true
+		}
+		return false
+	}
+
+	s1 := kRandomMembers(3, nodes, filterFunc)
+	s2 := kRandomMembers(3, nodes, filterFunc)
+	s3 := kRandomMembers(3, nodes, filterFunc)
+
+	if reflect.DeepEqual(s1, s2) {
+		t.Fatalf("unexpected equal")
+	}
+	if reflect.DeepEqual(s1, s3) {
+		t.Fatalf("unexpected equal")
+	}
+	if reflect.DeepEqual(s2, s3) {
+		t.Fatalf("unexpected equal")
+	}
+
+	for _, s := range [][]Member{s1, s2, s3} {
+		if len(s) != 3 {
+			t.Fatalf("bad len")
+		}
+		for _, m := range s {
+			if m.Name == "test0" {
+				t.Fatalf("Bad name")
+			}
+			if m.Status != StatusAlive {
+				t.Fatalf("Bad state")
+			}
+		}
 	}
 }
